Move the IContactAndReservationSettings to Integrations namespace
