--- src/Altinn.Profile.Integrations/Repositories/ProfileRepository.cs
@@ -1,134 +1,0 @@
-﻿#nullable enable
-
-using Altinn.Profile.Core.Domain;
-
-namespace Altinn.Profile.Integrations.Repositories;
-
-/// <summary>
-/// Generic repository for handling data access operations.
-/// </summary>
-/// <typeparam name="T">The type of the entity.</typeparam>
-/// <seealso cref="IRepository{T}" />
-internal class ProfileRepository<T> : IRepository<T>
-    where T : class
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="ProfileRepository{T}"/> class.
-    /// </summary>
-    internal ProfileRepository()
-    {
-    }
-
-    /// <summary>
-    /// Asynchronously adds a new entity to the database.
-    /// </summary>
-    /// <param name="entity">The entity to add.</param>
-    /// <returns>
-    /// A task that represents the asynchronous operation. The task result contains the added entity.
-    /// </returns>
-    public Task<T> AddAsync(T entity)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Adds multiple entities to the database asynchronously.
-    /// </summary>
-    /// <param name="entities">The entities to add.</param>
-    /// <returns>A task that represents the asynchronous operation.</returns>
-    public Task AddRangeAsync(IEnumerable<T> entities)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Deletes an entity from the database asynchronously based on its identifier.
-    /// </summary>
-    /// <param name="id">The identifier of the entity to delete.</param>
-    /// <returns>A task that represents the asynchronous operation.</returns>
-    public Task DeleteAsync(string id)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Deletes multiple entities from the database asynchronously based on their identifiers.
-    /// </summary>
-    /// <param name="ids">The identifiers of the entities to delete.</param>
-    /// <returns>A task that represents the asynchronous operation.</returns>
-    public Task DeleteRangeAsync(IEnumerable<string> ids)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Checks whether an entity exists asynchronously based on its identifier.
-    /// </summary>
-    /// <param name="id">The identifier of the entity.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains true if the entity exists, otherwise false.</returns>
-    public Task<bool> ExistsAsync(string id)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Retrieves all entities from the database asynchronously.
-    /// </summary>
-    /// <returns>A task that represents the asynchronous operation. The task result contains a collection of all entities.</returns>
-    public Task<IEnumerable<T>> GetAllAsync()
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Retrieves entities based on a filter, with optional sorting, pagination, and filtering.
-    /// </summary>
-    /// <param name="filter">A function to filter the entities.</param>
-    /// <param name="orderBy">A function to order the entities.</param>
-    /// <param name="skip">Number of entities to skip for pagination.</param>
-    /// <param name="take">Number of entities to take for pagination.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains a collection of entities matching the criteria.</returns>
-    public Task<IEnumerable<T>> GetAsync(Func<T, bool>? filter = null, Func<IEnumerable<T>, IOrderedEnumerable<T>>? orderBy = null, int? skip = null, int? take = null)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Retrieves an entity asynchronously based on its identifier.
-    /// </summary>
-    /// <param name="id">The identifier of the entity to retrieve.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains the entity if found, otherwise null.</returns>
-    public Task<T?> GetByIdAsync(string id)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Saves the changes made in the context asynchronously.
-    /// </summary>
-    /// <returns>A task that represents the asynchronous operation. The task result contains the number of state entries written to the database.</returns>
-    public Task<int> SaveChangesAsync()
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Updates an entity in the database asynchronously.
-    /// </summary>
-    /// <param name="entity">The entity to update.</param>
-    /// <returns>A task that represents the asynchronous operation.</returns>
-    public Task UpdateAsync(T entity)
-    {
-        throw new NotImplementedException();
-    }
-
-    /// <summary>
-    /// Updates multiple entities in the database asynchronously.
-    /// </summary>
-    /// <param name="entities">The entities to update.</param>
-    /// <returns>A task that represents the asynchronous operation.</returns>
-    public Task UpdateRangeAsync(IEnumerable<T> entities)
-    {
-        throw new NotImplementedException();
-    }
-}
